Move the auto-retry function to opto.utils
